pyproject.toml
[project]
name = "qa-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.4.2",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langfuse>=2.0.0",
    "openinference-instrumentation-openai-agents>=0.1.3",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "nest-asyncio>=1.6.0",
]
